TLDR: split on newlines rather than reading as an array
A single `socket.write` does not always result in a single
`socket.on('data')` event being triggered. The way this has manifested
in the past has been by multiple messages being concatenated into a
single message. To that end, the current strategy is to join messages
into an array by appending `,` on the sending side and transforming that
on the receiving side.

However, a new, reproducible case has popped up where the inverse is
happening: a single write resulting in 2 separate chunks of data coming
in. This isn't really something we can handle with the current strategy.

Turns out that this is essentially a solved problem: the `split` package
exports a `split()` function that a socket can be piped through and
which does all the bookkeeping to split chunks into proper messages.

So the new strategy is `append a newline`, and `split on newlines`. All
in all, this looks more like plain old json-lines.