Only care about tests/ if no globs given
We run our tests via `elm-test 'frontend/**/*Tests.elm'` as our app is a
big monorepo and we don't want our Elm tests in the top level `tests`
directory.

However, doing this gives us an error that the `tests/` folder doesn't
exist. I updated the code to only check for and use `tests/` if not
given a glob.