Fix bug adding tests/ to source-directories
This bug meant that running `elm-test "/not/default/path/*.elm"` would
error if `tests/` directory didn't exist, because we'd add it to the
`source-directories` of the new `elm.json` file we create.

This fix now ensures we only add the `source-directories` if:

- we're a package, and therefore using `tests/` for our tests
- we're called without a glob, and therefore are defaulting to `tests/`
- the `tests/` folder exists on disk. This covers the case where I call
`elm-test "tests/Foo.elm"` to only run one file.

If none of the above is true, we don't default to adding the `tests`
directory.