Prevent many small highlighted blocks in diff.
By comparing the number of edges of highlighted blocks to the number
of non-highlighted characters (highlighting is the result of
diffing),  messy diffs with lots of small highlighted blocks are
disabled.

Diffs will only be displayed if there are more plain characters than there
are edges highlighted blocks in both the actual and expected strings.

Fixes https://github.com/rtfeldman/node-test-runner/issues/263.