fix tests that check exit code of spawned elm test
Currently, a few tests check the `code` property of spawned processes.
However, the cross-spawn library does not define a `code` property on
the objects it returns; instead it defines a `status` property!

The test runner's unit tests where checking that the `code` was not
zero. Of course the code was not zero: it was `undefined` :angry:

As well as fixing the unit tests to check `.status` instead of `.code`
this commit fixes a unit test checked (or more accurately _did not
check_) the exit code of `elm-test --help` in light of
https://github.com/rtfeldman/node-test-runner/issues/359.

A prime example of why static typing is a boon for programming languages.