drop dependancy on find-parent-diff
Whilst this may look like it breaks elm-test when we run from a subdirectory
of the current project, it does not. elm-test is already broken when run
from a subdirectory. This just removes code that is broken and does not work.
If anyone wants to fix this reverting this commit would not be a bad place to
start.

In the meantime, one less dependancy is a good thing. All the more so becase
https://github.com/thlorenz/find-parent-dir has not been updated for 7 years.