Fixes (#448)
A cherry-pick of #442. The idea is let the little fixes from that PR be merged into master earlier and to reduce the diff in #442 to make that PR easier to review.

This PR improves the test suite, re-enables all the tests, makes sure we never disable tests again and improves our flow annotations.