Leverage `elm-json solve` (updated) (#451)
Fixes #328, fixes #277, fixes #273, fixes #323, fixes #374

zwilias  recently released [`elm-json`](https://github.com/zwilias/elm-json), a tool for interacting with `elm.json` files. One of the available subcommands is `solve`: Given an elm.json file (for a package or application), it tries to generate a single set of direct and indirect dependencies that satisfy all constraints. It has 2 flags that make it useful for `node-test-runner`:
- `--test` => Includes the `test-dependencies` and promotes them to direct dependencies
- `--extra` => Allows providing an extra set of packages that need to be direct dependencies of the final result

Since `elm-json` can be installed through `npm` and has binaries available for OS X, Windows and Linux (statically linked), it seems like a fine candidate to use in `node-test-runner` for calculating the final set of dependencies needed for the generated application.