Fix watch for packages when there are compilation errors in src/ (#453)
Fixes #399.

Previously, if there was an error in src/ for a package project, elm-test --watch would immediately exit. This was because we compile src/ first to get .elmi files for the files in there so that elmi-to-json can find the exact dependencies there. But since #451 we don’t use elmi-to-json to find exact dependencies anymore – we use elm-json instead – so there’s no need for that compilation step anymore. I removed that, which fixes the issue and also improves test run time since the compiler is invoked one time less.

As for the undefined log issue, that was because of promises being rejected with strings while the .catch code expected Error objects. "some string".message is undefined, so that’s where it came from. I updated to reject with Error objects instead.

This allowed for some pretty nice cleanups.