Fix "no tests found" errors for json and junit (#496)
Fixes https://github.com/rtfeldman/node-test-runner/issues/495

I consider this something that _should_ be a compile time error. It isn’t currently because we run the JS to figure out which exposed values are of type Test. But it used to be in 0.19.1-revision4 and older.

All compile time errors are printed human readable to stderr. This PR makes sure that happens for this error as well. There was already code in place for this, but it was only for the `console` reporter. I changed it to apply to all reporters.