Drop Node.js 10
Node.js 10 became end-of-life 2021-04-30.

There is a pull request that needs `worker_threads`:
https://github.com/rtfeldman/node-test-runner/pull/558.

The `worker_threads` module technically exists in Node.js 10, but is
behind the `--experimental-worker`. In Node.js 12 the module is marked
as stable and can be used without flags.

Since Node.js 10 has been end-of-life for many months now and we need
newer features, it’s time to drop support for Node.js 10.

The new lowest supported version is Node.js 12.13.0, which is the oldest
long term support release of Node.js 12: https://nodejs.org/es/blog/release/v12.13.0/

This PR:

- Updates the minium Node.js version to 12.13.0.
- Drops Node.js 10 from CI.
- Drops Node.js 15 from CI (since odd version numbers have short
  support, and Node.js 15 is end of life), and adds Node.js 16 (newest
  long term support) and Node.js 17 (current latest version).
- Resolves “we can do _this_ instead when Node.js 10 is dropped”
  comments in the code.