Allow both v1 and v2 of elm-solve-deps-wasm
This is commit switches to the same version constraint of
elm-solve-deps-wasm as elm-review uses these days. (Since 2.13.0, via
https://github.com/jfmengels/node-elm-review/pull/195.)

The benefit of this
commit is that if you install the latest version of both elm-test and
elm-review, you only get one copy of elm-solve-deps-wasm in
node_modules/ instead of two, saving about 288 KB.

Closes #636. This commit does the same thing as that PR, but without
updating the package-lock.json version. That will be done some other
time.